Fix many compiler warnings for gcc -Wextra for ogr, fmts and alg (https://trac.osgeo.org/gdal/ticket/5414)
git-svn-id: https://svn.osgeo.org/gdal/branches/1.11@27729 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/libgeotiff/geo_print.c
@@ -180,7 +180,7 @@ static void PrintKey(GeoKey *key, GTIFPrintMethod print, void *aux)
                   message[out_char++] = ch;
 
               /* flush message if buffer full */
-              if( out_char >= sizeof(message)-3 )
+              if( (size_t)out_char >= sizeof(message)-3 )
               {
                   message[out_char] = '\0';
                   print(message,aux);
